+++ HISTORY.rst
@@ -6,4 +6,9 @@ History
 0.0.6 (unreleased)
 ++++++++++++++++++
 
+- Initial possible working commit of incendiary
 - BREAKING: changed tracing name to be the same as hostname
+- Capture functionality
+- Task factories/wrappers
+- Sampling Class
+- Interservice tracing